Update Programme 9 Sept.py

+++ Assignment 1 /Programme 9 Sept.py
@@ -3,9 +3,10 @@
 #Programme Enter Weight in kg and convert into pounds
 
 weight = input('wnter a weight: ')
-weight = int(weight)
+weight = float(weight)
 
 pound = weight*2.204
-pound = int(pound)
+pound = float(pound)
 
 print('There are', pound ,'pounds in', weight ,'Kilograms')
+
